+++ examples/com/splunk/examples/index/Program.java
@@ -20,6 +20,7 @@ import com.splunk.EntityCollection;
 import com.splunk.Index;
 import com.splunk.Service;
 import com.splunk.Command;
+import com.splunk.SplunkException;
 
 public class Program {
     private static void list(Service service) {
@@ -64,8 +65,18 @@ public class Program {
         if (index == null)
             Command.error("Index '" + name + "' does not exists");
 
-        if (action.equals("clean"))
-            index.clean(180); // Timeout after 3 minutes.
+        if (action.equals("clean")) {
+            try {
+                index.clean(180);   // Timeout after 3 minutes.
+            } catch (SplunkException e) {
+                if (e.getCode() == SplunkException.INTERRUPTED) {
+                    // User pressed Ctrl-C
+                    return;
+                } else {
+                    throw e;
+                }
+            } 
+        }
         else if (action.equals("disable"))
             index.disable();
         else if (action.equals("enable"))