+++ CHANGELOG.md
@@ -74,6 +74,10 @@
 * Index.setAssureUTF8() fails for Splunk >= 5.0, since that field has become a
   global setting instead of a per-index setting.
 
+* Index.clean() now throws SplunkException.INTERRUPTED when interrupted instead
+  of swallowing the condition. Additionally the timeout parameter is obeyed more
+  accurately.
+
 * WindowsRegistryInput.getType() and setType() changed to be String[] instead of String.
 
 * DistributedPeer.getBuild() now returns an int instead of a String to be